brought the about page to main
