+++ client/pages/index.js
@@ -1,129 +1,37 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import { Inter } from '@next/font/google'
-import styles from '@/styles/Home.module.css'
+import { ChakraProvider } from '@chakra-ui/react';
+import styles from '@/styles/Home.module.css';
+import Navbar from '../src/components/Navbar';
+import Image from 'next/image';
 
-const inter = Inter({ subsets: ['latin'] })
-
-export default function Home() {
+const Page = () => {
     return (
-        <>
-            <Head>
-                <title>Hanawilo Resource Library</title>
-                <meta name='description' content='Generated by create next app' />
-                <meta
-                    name='viewport'
-                    content='width=device-width, initial-scale=1'
-                />
-                <link rel='icon' href='/favicon.ico' />
-            </Head>
-            <main className={styles.main}>
-                <div className={styles.description}>
-                    <p>
-                        Get started by editing&nbsp;
-                        <code className={styles.code}>src/pages/index.js</code>
-                    </p>
-                    <div>
-                        <a
-                            href='https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app'
-                            target='_blank'
-                            rel='noopener noreferrer'
-                        >
-                            By{' '}
-                            <Image
-                                src='/vercel.svg'
-                                alt='Vercel Logo'
-                                className={styles.vercelLogo}
-                                width={100}
-                                height={24}
-                                priority
-                            />
-                        </a>
-                    </div>
-                </div>
-
+        <ChakraProvider>
+            <Navbar />
+            <main>
+                <h1>Our Hanawilo Resource Library Project</h1>
                 <div className={styles.center}>
                     <Image
-                        className={styles.logo}
-                        src='/next.svg'
-                        alt='Next.js Logo'
-                        width={180}
-                        height={37}
-                        priority
+                        src='/logo.svg'
+                        alt='Hanawilo Logo'
+                        className={styles.vercelLogo}
+                        width={100}
+                        height={24}
                     />
+                </div>
+                <div className={styles.center}>
                     <div className={styles.thirteen}>
                         <Image
-                            src='/thirteen.svg'
-                            alt='13'
-                            width={40}
-                            height={31}
-                            priority
+                            src='/next.svg'
+                            alt='Next.js Logo'
+                            className={styles.logo}
+                            width={180}
+                            height={37}
                         />
                     </div>
                 </div>
-
-                <div className={styles.grid}>
-                    <a
-                        href='https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app'
-                        className={styles.card}
-                        target='_blank'
-                        rel='noopener noreferrer'
-                    >
-                        <h2 className={inter.className}>
-                            Docs <span>-&gt;</span>
-                        </h2>
-                        <p className={inter.className}>
-                            Find in-depth information about Next.js features
-                            and&nbsp;API.
-                        </p>
-                    </a>
-
-                    <a
-                        href='https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app'
-                        className={styles.card}
-                        target='_blank'
-                        rel='noopener noreferrer'
-                    >
-                        <h2 className={inter.className}>
-                            Learn <span>-&gt;</span>
-                        </h2>
-                        <p className={inter.className}>
-                            Learn about Next.js in an interactive course
-                            with&nbsp;quizzes!
-                        </p>
-                    </a>
-
-                    <a
-                        href='https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app'
-                        className={styles.card}
-                        target='_blank'
-                        rel='noopener noreferrer'
-                    >
-                        <h2 className={inter.className}>
-                            Templates <span>-&gt;</span>
-                        </h2>
-                        <p className={inter.className}>
-                            Discover and deploy boilerplate example
-                            Next.js&nbsp;projects.
-                        </p>
-                    </a>
-
-                    <a
-                        href='https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app'
-                        className={styles.card}
-                        target='_blank'
-                        rel='noopener noreferrer'
-                    >
-                        <h2 className={inter.className}>
-                            Deploy <span>-&gt;</span>
-                        </h2>
-                        <p className={inter.className}>
-                            Instantly deploy your Next.js site to a shareable URL
-                            with&nbsp;Vercel.
-                        </p>
-                    </a>
-                </div>
             </main>
-        </>
+        </ChakraProvider>
     );
-}
+};
+
+export default Page;
